Revert "Add PollingQueueInputStream and TakingQueueInputStream to replace"
This reverts commit b21c2ac4009a31def29055468e05f290b1019bb5.
